requirements.txt
pandas==1.3.2
joblib==1.0.1
Flask==2.0.1
expertai_nlapi==2.3.1
numpy==1.21.2
plotly==5.2.2
Werkzeug==2.0.1
Flask_RESTful==0.3.9
nltk==3.6.2
python-dotenv==0.19.0
scikit_learn==0.24.2
gunicorn
